actualizacion 0ctubre 2025

+++ src/main/java/com/ipn/mx/demowebsocket/basedatos/domain/repository/PuntajeDetalleRepository.java
@@ -3,5 +3,4 @@ package com.ipn.mx.demowebsocket.basedatos.domain.repository;
 import com.ipn.mx.demowebsocket.basedatos.domain.entity.PuntajeDetalle;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PuntajeDetalleRepository extends JpaRepository<PuntajeDetalle, Integer> {
-}
+public interface PuntajeDetalleRepository extends JpaRepository<PuntajeDetalle, Long> {}
